@@ -9,22 +9,22 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^17.3.0-rc",
-    "@angular/cdk": "^17.3.0-rc",
-    "@angular/common": "^17.3.0-rc",
-    "@angular/compiler": "^17.3.0-rc",
-    "@angular/core": "^17.3.0-rc",
-    "@angular/forms": "^17.3.0-rc",
-    "@angular/material": "^17.3.0-rc",
-    "@angular/platform-browser": "^17.3.0-rc",
+    "@angular/animations": "^18.0.0",
+    "@angular/cdk": "^18.0.0",
+    "@angular/common": "^18.0.0",
+    "@angular/compiler": "^18.0.0",
+    "@angular/core": "^18.0.0",
+    "@angular/forms": "^18.0.0",
+    "@angular/material": "^18.0.0",
+    "@angular/platform-browser": "^18.0.0",
     "rxjs": "~7.8.0",
     "tslib": "^2.3.0",
     "zone.js": "~0.14.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^17.3.0-rc",
-    "@angular/cli": "^17.3.0-rc",
-    "@angular/compiler-cli": "^17.3.0-rc",
+    "@angular/build": "^18.0.1",
+    "@angular/cli": "^18.0.0",
+    "@angular/compiler-cli": "^18.0.0",
     "typescript": "~5.4.0"
   }
 }
