@@ -14,6 +14,7 @@ import {
   PLATFORM_ID,
   computed,
   inject,
+  signal,
 } from '@angular/core';
 import {takeUntilDestroyed, toObservable} from '@angular/core/rxjs-interop';
 import {
@@ -64,19 +65,19 @@ export class SecondaryNavigation implements OnInit {
   private readonly router = inject(Router);
   private readonly window = inject(WINDOW);
 
-  isSecondaryNavVisible = this.navigationState.isMobileNavVisible;
-  primaryActiveRouteItem = this.navigationState.primaryActiveRouteItem;
-  maxVisibleLevelsOnSecondaryNav = computed(() =>
+  readonly isSecondaryNavVisible = this.navigationState.isMobileNavVisible;
+  readonly primaryActiveRouteItem = this.navigationState.primaryActiveRouteItem;
+  readonly maxVisibleLevelsOnSecondaryNav = computed(() =>
     this.primaryActiveRouteItem() === PagePrefix.REFERENCE ? 1 : 2,
   );
-  navigationItemsSlides = this.navigationState.expandedItems;
+  readonly navigationItemsSlides = this.navigationState.expandedItems;
   navigationItems: NavigationItem[] | undefined;
 
   translateX = computed(() => {
     const level = this.navigationState.expandedItems()?.length ?? 0;
     return `translateX(${-level * 100}%)`;
   });
-  transition: string = '0ms';
+  transition = signal('0ms');
 
   readonly PRIMARY_NAV_ID = PRIMARY_NAV_ID;
   readonly SECONDARY_NAV_ID = SECONDARY_NAV_ID;
@@ -182,7 +183,7 @@ export class SecondaryNavigation implements OnInit {
       return;
     }
     setTimeout(() => {
-      this.transition = `${ANIMATION_DURATION}ms`;
+      this.transition.set(`${ANIMATION_DURATION}ms`);
     }, ANIMATION_DURATION);
   }
 
