@@ -6,8 +6,9 @@
  * found in the LICENSE file at https://angular.dev/license
  */
 
-import {Injectable, VERSION, computed, signal} from '@angular/core';
+import {Injectable, VERSION, computed, inject, signal} from '@angular/core';
 import {VERSIONS_CONFIG} from '../constants/versions';
+import {WINDOW} from '@angular/docs';
 
 export interface Version {
   displayName: string;
@@ -17,39 +18,79 @@ export interface Version {
 
 export type VersionMode = 'stable' | 'deprecated' | 'rc' | 'next' | number;
 
-export const INITIAL_DOCS_VERSION = 17;
-export const VERSION_PATTERN_PLACEHOLDER = '{{version}}';
+export const INITIAL_ADEV_DOCS_VERSION = 18;
+export const VERSION_PLACEHOLDER = '{{version}}';
+export const MODE_PLACEHOLDER = '{{prefix}}';
 
 @Injectable({
   providedIn: 'root',
 })
 export class VersionManager {
+  private readonly window = inject(WINDOW);
+
+  // Note: We can assume that if the URL starts with v{{version}}, it is documentation for previous versions of Angular.
+  // Based on URL we can indicate as well if it's rc or next Docs version.
+  private get currentVersionMode(): VersionMode {
+    const hostname = this.window.location.hostname;
+    if (hostname.startsWith('v')) return 'deprecated';
+    if (hostname.startsWith('rc')) return 'rc';
+    if (hostname.startsWith('next')) return 'next';
+
+    return 'stable';
+  }
+
   versions = signal<Version[]>([
-    ...VERSIONS_CONFIG.mainVersions.map((item) =>
-      this.mapToVersion(item as Pick<Version, 'url' | 'version'>),
-    ),
-    ...this.getHistoricalVersions(),
+    ...this.getRecentVersions(),
+    ...this.getAdevVersions(),
+    ...this.getAioVersions(),
   ]);
 
   currentDocsVersion = computed(() => {
     return this.versions().find(
-      (version) => version.version.toString() === VERSIONS_CONFIG.currentVersion,
+      (version) => version.version.toString() === this.currentVersionMode,
     );
   });
 
-  private getHistoricalVersions(): Version[] {
-    const historicalVersions: Version[] = [];
-    for (let version = Number(VERSION.major) - 1; version >= INITIAL_DOCS_VERSION; version--) {
-      historicalVersions.push({
-        url: VERSIONS_CONFIG.historicalVersionsLinkPattern.replace(
-          VERSION_PATTERN_PLACEHOLDER,
-          version.toString(),
-        ),
+  // List of Angular Docs versions which includes current version, next and rc.
+  private getRecentVersions(): Version[] {
+    return [
+      {
+        url: this.getAdevDocsUrl('next'),
+        displayName: `next`,
+        version: 'next',
+      },
+      // Note: 'rc' should not be visible for now
+      // {
+      //   url: this.getAdevDocsUrl('rc'),
+      //   displayName: `rc`,
+      //   version: 'rc',
+      // },
+      {
+        url: this.getAdevDocsUrl(Number(VERSION.major)),
+        displayName: this.getVersion(Number(VERSION.major)),
+        version: this.currentVersionMode,
+      },
+    ];
+  }
+
+  // List of Angular Docs versions hosted on angular.dev domain.
+  private getAdevVersions(): Version[] {
+    const adevVersions: Version[] = [];
+    for (let version = Number(VERSION.major) - 1; version >= INITIAL_ADEV_DOCS_VERSION; version--) {
+      adevVersions.push({
+        url: this.getAdevDocsUrl(version),
         displayName: this.getVersion(version),
-        version,
+        version: 'deprecated',
       });
     }
-    return historicalVersions;
+    return adevVersions;
+  }
+
+  // List of Angular Docs versions hosted on angular.io domain.
+  private getAioVersions(): Version[] {
+    return VERSIONS_CONFIG.aioVersions.map((item) =>
+      this.mapToVersion(item as Pick<Version, 'url' | 'version'>),
+    );
   }
 
   private mapToVersion(value: Pick<Version, 'url' | 'version'>): Version {
@@ -60,14 +101,23 @@ export class VersionManager {
   }
 
   private getVersion(versionMode: VersionMode): string {
-    if (versionMode === 'stable') {
-      return 'v17';
-      // Temporarily commenting out till this works correctly
-      // return `v${VERSION.major}`;
+    if (versionMode === 'stable' || versionMode === 'deprecated') {
+      return `v${VERSION.major}`;
     }
     if (Number.isInteger(versionMode)) {
       return `v${versionMode}`;
     }
     return versionMode.toString();
+  }
+
+  private getAdevDocsUrl(version: VersionMode): string {
+    const docsUrlPrefix = isNaN(Number(version)) ? `` : 'v';
+
+    return VERSIONS_CONFIG.aDevVersionsLinkPattern
+      .replace(MODE_PLACEHOLDER, docsUrlPrefix)
+      .replace(
+        VERSION_PLACEHOLDER,
+        `${version.toString() === 'stable' ? '' : `${version.toString()}.`}`,
+      );
   }
 }
