@@ -64,7 +64,7 @@ describe('NgModule', () => {
       expect(() => createInjector(AModule)).toThrowError(
         'NG0200: Circular dependency in DI detected for AModule. ' +
           'Dependency path: AModule > BModule > AModule. ' +
-          'Find more at https://angular.io/errors/NG0200',
+          'Find more at https://angular.dev/errors/NG0200',
       );
     });
 
