@@ -10,7 +10,7 @@ import {ErrorCode} from './error_code';
 
 /**
  * Contains a set of error messages that have detailed guides at angular.io.
- * Full list of available error guides can be found at https://angular.io/errors
+ * Full list of available error guides can be found at https://angular.dev/errors
  */
 export const COMPILER_ERRORS_WITH_GUIDES = new Set([
   ErrorCode.DECORATOR_ARG_NOT_LITERAL,
