@@ -6,7 +6,15 @@
  * found in the LICENSE file at https://angular.io/license
  */
 
-import {AST, Interpolation, PropertyRead, TmplAstNode} from '@angular/compiler';
+import {
+  AST,
+  ASTWithSource,
+  BindingType,
+  Interpolation,
+  PropertyRead,
+  TmplAstBoundAttribute,
+  TmplAstNode,
+} from '@angular/compiler';
 import ts from 'typescript';
 
 import {ErrorCode, ExtendedTemplateDiagnosticName} from '../../../../diagnostics';
@@ -34,10 +42,36 @@ class InterpolatedSignalCheck extends TemplateCheckWithVisitor<ErrorCode.INTERPO
     component: ts.ClassDeclaration,
     node: TmplAstNode | AST,
   ): NgTemplateDiagnostic<ErrorCode.INTERPOLATED_SIGNAL_NOT_INVOKED>[] {
+    // interpolations like `{{ mySignal }}`
     if (node instanceof Interpolation) {
       return node.expressions
         .filter((item): item is PropertyRead => item instanceof PropertyRead)
         .flatMap((item) => buildDiagnosticForSignal(ctx, item, component));
+    }
+    // bound properties like `[prop]="mySignal"`
+    else if (node instanceof TmplAstBoundAttribute) {
+      const symbol = ctx.templateTypeChecker.getSymbolOfNode(node, component);
+      // we skip the check if the node is an input binding
+      if (symbol !== null && symbol.kind === SymbolKind.Input) {
+        return [];
+      }
+      // otherwise, we check if the node is
+      if (
+        // a bound property like `[prop]="mySignal"`
+        (node.type === BindingType.Property ||
+          // or a class binding like `[class.myClass]="mySignal"`
+          node.type === BindingType.Class ||
+          // or a style binding like `[style.width]="mySignal"`
+          node.type === BindingType.Style ||
+          // or an attribute binding like `[attr.role]="mySignal"`
+          node.type === BindingType.Attribute ||
+          // or an animation binding like `[@myAnimation]="mySignal"`
+          node.type === BindingType.Animation) &&
+        node.value instanceof ASTWithSource &&
+        node.value.ast instanceof PropertyRead
+      ) {
+        return buildDiagnosticForSignal(ctx, node.value.ast, component);
+      }
     }
     return [];
   }
