@@ -13,6 +13,8 @@ import {
   EventContractContainer,
   EventInfoWrapper,
   registerDispatcher,
+  isSupportedEvent,
+  isCaptureEvent,
 } from '@angular/core/primitives/event-dispatch';
 
 import {APP_BOOTSTRAP_LISTENER, ApplicationRef, whenStable} from '../application/application_ref';
@@ -22,7 +24,7 @@ import {inject} from '../di/injector_compatibility';
 import {Provider} from '../di/interface/provider';
 import {setDisableEventReplayImpl} from '../render3/instructions/listener';
 import {TNode, TNodeType} from '../render3/interfaces/node';
-import {RNode} from '../render3/interfaces/renderer_dom';
+import {RElement, RNode} from '../render3/interfaces/renderer_dom';
 import {CLEANUP, LView, TView} from '../render3/interfaces/view';
 import {isPlatformBrowser} from '../render3/util/misc_utils';
 import {unwrapRNode} from '../render3/util/view_utils';
@@ -42,6 +44,11 @@ function getJsactionData(container: EarlyJsactionDataContainer) {
 }
 
 const JSACTION_ATTRIBUTE = 'jsaction';
+
+/**
+ * Associates a DOM element with `jsaction` attribute to a map that contains info about all event
+ * types (event names) and corresponding listeners.
+ */
 const jsactionMap: Map<Element, Map<string, Function[]>> = new Map();
 
 /**
@@ -57,18 +64,19 @@ export function withEventReplay(): Provider[] {
     {
       provide: ENVIRONMENT_INITIALIZER,
       useValue: () => {
-        setDisableEventReplayImpl((rEl, eventName, listenerFn) => {
+        setDisableEventReplayImpl((rEl: RElement, eventName: string, listenerFn: VoidFunction) => {
           if (rEl.hasAttribute(JSACTION_ATTRIBUTE)) {
-            const el = unwrapRNode(rEl) as Element;
+            const el = rEl as unknown as Element;
             // We don't immediately remove the attribute here because
             // we need it for replay that happens after hydration.
             if (!jsactionMap.has(el)) {
               jsactionMap.set(el, new Map());
             }
-            if (!jsactionMap.get(el)!.has(eventName)) {
-              jsactionMap.get(el)!.set(eventName, []);
+            const eventMap = jsactionMap.get(el)!;
+            if (!eventMap.has(eventName)) {
+              eventMap.set(eventName, []);
             }
-            jsactionMap.get(el)!.get(eventName)!.push(listenerFn);
+            eventMap.get(eventName)!.push(listenerFn);
           }
         });
       },
@@ -115,6 +123,9 @@ export function withEventReplay(): Provider[] {
                 for (const el of jsactionMap.keys()) {
                   el.removeAttribute(JSACTION_ATTRIBUTE);
                 }
+                // After hydration, we shouldn't need to do anymore work related to
+                // event replay anymore.
+                setDisableEventReplayImpl(() => {});
               }
             });
           };
@@ -133,7 +144,7 @@ export function withEventReplay(): Provider[] {
 export function collectDomEventsInfo(
   tView: TView,
   lView: LView,
-  eventTypesToReplay: Set<string>,
+  eventTypesToReplay: {regular: Set<string>; capture: Set<string>},
 ): Map<Element, string[]> {
   const events = new Map<Element, string[]>();
   const lCleanup = lView[CLEANUP];
@@ -148,15 +159,14 @@ export function collectDomEventsInfo(
       continue;
     }
     const name: string = firstParam;
-    if (
-      name === 'mouseenter' ||
-      name === 'mouseleave' ||
-      name === 'pointerenter' ||
-      name === 'pointerleave'
-    ) {
+    if (!isSupportedEvent(name)) {
       continue;
     }
-    eventTypesToReplay.add(name);
+    if (isCaptureEvent(name)) {
+      eventTypesToReplay.capture.add(name);
+    } else {
+      eventTypesToReplay.regular.add(name);
+    }
     const listenerElement = unwrapRNode(lView[secondParam]) as any as Element;
     i++; // move the cursor to the next position (location of the listener idx)
     const useCaptureOrIndx = tCleanup[i++];
@@ -192,7 +202,7 @@ export function setJSActionAttribute(
 }
 
 function handleEvent(event: EventInfoWrapper) {
-  const nativeElement = unwrapRNode(event.getAction()!.element) as Element;
+  const nativeElement = event.getAction()!.element as Element;
   const handlerFns = jsactionMap.get(nativeElement)?.get(event.getEventType());
   if (!handlerFns) {
     return;
