@@ -50,7 +50,7 @@ describe('di', () => {
         expect(() => {
           TestBed.createComponent(App).detectChanges();
         }).toThrowError(
-          'NG0201: No provider for DirNotOnSelf found in NodeInjector. Find more at https://angular.io/errors/NG0201',
+          'NG0201: No provider for DirNotOnSelf found in NodeInjector. Find more at https://angular.dev/errors/NG0201',
         );
       });
     });
