@@ -5803,7 +5803,7 @@ describe('reactive forms integration tests', () => {
       expect(() => {
         fixture.detectChanges();
       }).toThrowError(
-        `NG01203: No value accessor for form control name: 'control'. Find more at https://angular.io/errors/NG01203`,
+        `NG01203: No value accessor for form control name: 'control'. Find more at https://angular.dev/errors/NG01203`,
       );
 
       // Making sure that cleanup between tests doesn't cause any issues
